Refactoring Dagger. MainModule.kt deleted. Work with inject constructors, some binds for interfaces added.
