+++ app/src/main/java/com/example/users/ui/MainActivity.kt
@@ -1,7 +1,8 @@
-package com.example.users
+package com.example.users.ui
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import com.example.users.R
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
