+++ app/src/main/res/navigation/main_nav_graph.xml
@@ -6,6 +6,6 @@
 
     <fragment
         android:id="@+id/mainFragment"
-        android:name="com.example.users.usersfragment.MainFragment"
+        android:name="com.example.users.ui.MainFragment"
         android:label="MainFragment" />
 </navigation>
